+++ src/GridCal/Engine/Simulations/ContinuationPowerFlow/continuation_power_flow.py
@@ -988,7 +988,7 @@ if __name__ == '__main__':
 
     # just for this test
     numeric_circuit = main_circuit.compile()
-    numeric_inputs = numeric_circuit.compute()
+    numeric_inputs = numeric_circuit.compute(ignore_single_node_islands=options.ignore_single_node_islands)
     Sbase = zeros(len(main_circuit.buses), dtype=complex)
     Vbase = zeros(len(main_circuit.buses), dtype=complex)
     for c in numeric_inputs: