+++ src/GridCal/Engine/Simulations/Stochastic/lhs_driver.py
@@ -82,7 +82,8 @@ class LatinHypercubeSampling(QThread):
         # compile
         # print('Compiling...', end='')
         numerical_circuit = self.circuit.compile()
-        numerical_islands = numerical_circuit.compute(branch_tolerance_mode=self.options.branch_impedance_tolerance_mode)
+        numerical_islands = numerical_circuit.compute(branch_tolerance_mode=self.options.branch_impedance_tolerance_mode,
+                                                      ignore_single_node_islands=self.options.ignore_single_node_islands)
 
         lhs_results.bus_types = numerical_circuit.bus_types
 
@@ -202,7 +203,8 @@ class LatinHypercubeSampling(QThread):
 
         # compile the numerical circuit
         numerical_circuit = self.circuit.compile()
-        numerical_input_islands = numerical_circuit.compute(branch_tolerance_mode=self.options.branch_impedance_tolerance_mode)
+        numerical_input_islands = numerical_circuit.compute(branch_tolerance_mode=self.options.branch_impedance_tolerance_mode,
+                                                            ignore_single_node_islands=self.options.ignore_single_node_islands)
 
         max_iter = batch_size * len(numerical_input_islands)
         Sbase = numerical_circuit.Sbase